Move test, skip when no Ractor
